+++ src/yb/util/slice.h
@@ -200,7 +200,7 @@ inline std::ostream& operator<<(std::ostream& o, const Slice& s) {
 }
 
 inline int Slice::compare(const Slice& b) const {
-  const int min_len = (size_ < b.size_) ? size_ : b.size_;
+  const size_t min_len = (size_ < b.size_) ? size_ : b.size_;
   int r = MemCompare(data_, b.data_, min_len);
   if (r == 0) {
     if (size_ < b.size_) r = -1;
